+++ gum/managers.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
+
 from gum.utils import elasticsearch_connection
 
 
@@ -8,9 +9,16 @@ class ElasticsearchManager(object):
     to Elasticsearch from a specific model.
     """
 
-    def __init__(self, model=None, mapping_type=None):
+    def __init__(self, model=None, mapping_type=None, urls=None):
         self.model = model
         self.mapping_type = mapping_type
+        self.urls = urls
+
+    def get_elasticsearch_connection(self):
+        """Gets the Elasticsearch connection with the urls attribute"""
+        if self.mapping_type is not None:
+            return self.mapping_type.get_elasticsearch_connection()
+        return elasticsearch_connection(urls=self.urls)
 
     def search(self, **kwargs):
         """Partial application of `search` function from Elasticsearch
@@ -18,7 +26,7 @@ class ElasticsearchManager(object):
 
         :param kwargs:
         """
-        es = elasticsearch_connection()
+        es = self.get_elasticsearch_connection()
         if 'index' not in kwargs:
             kwargs["index"] = self.mapping_type.index
         if 'doc_type' not in kwargs:
@@ -32,3 +40,20 @@ class ElasticsearchManager(object):
         :return:
         """
         return self.mapping_type.index_document(instance)
+
+
+class GenericElasticsearchManager(ElasticsearchManager):
+    """Generic Elasticsearch manager to make queries without using MappingTypes."""
+
+    def search(self, **kwargs):
+        """For this manager it's mandatory to specify index and doc_type on each call:
+
+        >>> elasticsearch = GenericElasticsearchManager()
+        >>> elasticsearch.search(index="index-name", doc_type="mapping-type-name")
+
+        :param kwargs:
+        :return:
+        """
+        assert "index" in kwargs
+        assert "doc_type" in kwargs
+        return super(GenericElasticsearchManager, self).search(**kwargs)